+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+###### February 6, 2021
+- Completed work for `Challenge Result`
+
 ###### February 2, 2021
 - Changes made to support backup gauge option
 - Fixed red laser end texture being slightly too short